class_4/pyproject.toml
[project]
name = "class-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.73.6",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
